feat: Updated code structure
+ Fixed lambda functions
+ Updated NumberValue to use double precision
+ Updated error messages to print line and column information
+ Updated BuiltInFunction so that it can support multiple running instances of the language
- Removed constant values

+++ src/me/senseiwells/arucas/nodes/NumberNode.java
@@ -11,7 +11,7 @@ public class NumberNode extends Node {
     
     public NumberNode(Token token) {
         super(token);
-        this.value = new NumberValue(Float.parseFloat(token.content));
+        this.value = new NumberValue(Double.parseDouble(token.content));
         this.value.setPos(this.startPos, this.endPos);
     }
